remove exceptions, they stopped working in webassembly. i'm not happy with the resulting code, but don't know a better way.

+++ nucleus/tile_scheduler/Scheduler.cpp
@@ -223,14 +223,14 @@ void Scheduler::purge_ram_cache()
 
 void Scheduler::persist_tiles()
 {
-    try {
-        const auto start = std::chrono::steady_clock::now();
-        m_ram_cache.write_to_disk(disk_cache_path());
-        const auto diff = std::chrono::steady_clock::now() - start;
-        if (diff > std::chrono::milliseconds(50))
-            fmt::println(stderr, "Scheduler::persist_tiles took {} for {} quads.", std::chrono::duration_cast<std::chrono::milliseconds>(diff), m_ram_cache.n_cached_objects());
-    } catch (const std::runtime_error& e) {
-        qDebug("Writing tiles to disk into %s failed: %s. Removing all files.", disk_cache_path().c_str(), e.what());
+    const auto start = std::chrono::steady_clock::now();
+    const auto r = m_ram_cache.write_to_disk(disk_cache_path());
+    const auto diff = std::chrono::steady_clock::now() - start;
+    if (diff > std::chrono::milliseconds(50))
+        fmt::println(stderr, "Scheduler::persist_tiles took {} for {} quads.", std::chrono::duration_cast<std::chrono::milliseconds>(diff), m_ram_cache.n_cached_objects());
+
+    if (!r.has_value()) {
+        qDebug("Writing tiles to disk into %s failed: %s. Removing all files.", disk_cache_path().c_str(), r.error().c_str());
         std::filesystem::remove_all(disk_cache_path());
     }
 }
@@ -267,11 +267,11 @@ void Scheduler::update_stats()
 
 void Scheduler::read_disk_cache()
 {
-    try {
-        m_ram_cache.read_from_disk(disk_cache_path());
+    const auto r = m_ram_cache.read_from_disk(disk_cache_path());
+    if (r.has_value()) {
         update_stats();
-    } catch (const std::runtime_error& e) {
-        qDebug("Reading tiles from disk cache (%s) failed: \n%s\nRemoving all files.", disk_cache_path().c_str(), e.what());
+    } else {
+        qDebug("Reading tiles from disk cache (%s) failed: \n%s\nRemoving all files.", disk_cache_path().c_str(), r.error().c_str());
         std::filesystem::remove_all(disk_cache_path());
     }
 }
